backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.43
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.18.4
pgvector==0.3.6
redis==5.0.8
celery[redis]==5.4.0
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.35
langsmith==0.4.38
dashscope==1.24.2
zhipuai==2.1.5.20250825
openai==1.109.1
python-docx==1.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-pptx==1.0.2
openpyxl==3.1.5
tiktoken==0.11.0
minio==7.2.18
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0
httpx==0.28.1
sse-starlette==3.0.2
pytest==8.4.2
pytest-asyncio==1.2.0
ruff==0.12.0
